Add files via upload
A different version of the code using "flags" and while-loops. The results aren't that great, but I think I'm onto something. Need to adjust values and filter out measurements over a certain distance and faulty measurements. 

NOTICE: Servo and DC motor are commented out in this version, in order to be able to just test the sensors without the car driving away.